Revert "moved some stuff around and created components for new pages"
This reverts commit 0029d4bae295f93fdd532a70cbe0f5fd87195d4e.
