+++ frontend/crd/src/app/dashboard/dashboard.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { DashboardComponent } from './dashboard.component';
 import {MockComponent} from "ng-mocks";
 import {EventsComponent} from "./events/events.component";
-import {MilestonesComponent} from "../milestones-page/milestones/milestones.component";
+import {MilestonesComponent} from "../portfolio/milestones/milestones.component";
 
 describe('DashboardComponent', () => {
   let component: DashboardComponent;
