+++ frontend/crd/src/app/dashboard/dashboard.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { DashboardComponent } from './dashboard.component';
 import { MatCardModule } from "@angular/material/card";
 import { EventsComponent } from './events/events.component';
-import {MilestonesModule} from "../milestones-page/milestones/milestones.module";
+import {PortfolioModule} from "../portfolio/portfolio.module";
 
 
 
@@ -17,7 +17,7 @@ import {MilestonesModule} from "../milestones-page/milestones/milestones.module"
   imports: [
     CommonModule,
     MatCardModule,
-    MilestonesModule
+    PortfolioModule
   ]
 })
 export class DashboardModule { }
