--- frontend/crd/src/app/milestones-page/milestones/milestones.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MatCardModule } from "@angular/material/card";
-import { MilestonesComponent } from './milestones.component';
-
-
-
-@NgModule({
-  declarations: [
-    MilestonesComponent,
-  ],
-  exports: [
-    MilestonesComponent
-  ],
-  imports: [
-    CommonModule,
-    MatCardModule,
-  ]
-})
-export class MilestonesModule { }
