+++ frontend/crd/src/app/portfolio/portfolio.module.ts
@@ -1,21 +1,22 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatCardModule } from "@angular/material/card";
+import {MilestonesComponent} from "./milestones/milestones.component";
 import {PortfolioComponent} from "./portfolio.component";
-import {MilestonesModule} from "../milestones-page/milestones/milestones.module";
 
 
 
 @NgModule({
   declarations: [
     PortfolioComponent,
+    MilestonesComponent,
   ],
   exports: [
+    MilestonesComponent
   ],
   imports: [
     CommonModule,
-    MatCardModule,
-    MilestonesModule
+    MatCardModule
   ]
 })
 export class PortfolioModule { }
